requirements.txt
selenium
pandas
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
oauth2client
Pillow
ipython
